third_party/blink/web_tests/fast/forms/week/week-picker-ax.html

<!DOCTYPE html>
<meta name=fuzzy content="maxDifference=0-3; totalPixels=0-1000">
  <script src="../../../resources/testharness.js"></script>
  <script src="../../../resources/testharnessreport.js"></script>
  <script src="../../../resources/testdriver.js"></script>
  <script src="../../../resources/testdriver-vendor.js"></script>
  <script src="../resources/picker-common.js"></script>
  <script src="../calendar-picker/resources/calendar-picker-common.js"></script>


<input type="week" id="week1" value="2000-W13">

<script>
promise_test(() => {
  let week1 = document.getElementById('week1');
  return openPicker(week1).then(() => {
    let markDirtyCounter = 0;

    let promise = new Promise((resolve, reject) => {
      accessibilityController.setNotificationListener(function(axnode, type) {
        if (type == 'MarkDirty' && ++ markDirtyCounter >= 2) {
          assert_equals(week1.value, "2000-W14", "Arrow key in week picker should have changed value");
          resolve();
        }
      });

      // In case of failure, reject the promise so we get some logs instead of timing out.
      window.setTimeout(() => {
        reject(`Didn't receive expected events.  Got ${markDirtyCounter} MarkDirty`);
      }, 1500);
    });

    eventSender.keyDown('ArrowDown');
    // This forces an a11y flush.
    accessibilityController.focusedElement;
    return promise;
  });
}, 'Tests if typing an arrow key dispatches |MarkDirty| a11y events.');

</script>
